Extract GetMaze querie to queries package
